lab 2

+++ labs/1/NumYears27.java
@@ -9,9 +9,9 @@ public class NumYears27 {
 		
 		System.out.print("Enter the number of minutes: ");
 		
-		int min = input.nextInt();
-		int years = min / 525600;
-		int days = (min % 525600) / 1440;
+		long min = input.nextLong();
+		long years = min / 525600;
+		long days = (min % 525600) / 1440;
 		
 		System.out.println(min + " minutes is approximately " + years + " years and " + days + " days");
